Install script now split into smaller files.

+++ install_scripts/fonts.sh
@@ -6,12 +6,23 @@ sudo yes Y | apt install \
   ttf-dejavu \
   fonts-powerline \
   fonts-font-awesome \
+  dtrx
 
-mkdir "$HOME/.fonts"
-cd "$HOME/.fonts"
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.0.0/SpaceMono.zip
-sudo yes Y | apt install dtrx
-dtrx SpaceMono.zip
-fc-cache -fv
-rm SpaceMono.zip
-cd
+fonts_dir="$HOME/.fonts"
+if [ ! -d "$fonts_dir" ]; then
+  mkdir "$HOME/.fonts"
+  else
+    echo ".fonts dir already exists"
+fi
+
+spacemono_dir="$HOME/.fonts/SpaceMono/"
+if [ ! -d "$spacemono_dir" ]; then
+  cd "$HOME/.fonts"
+  wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.0.0/SpaceMono.zip
+  dtrx SpaceMono.zip
+  fc-cache -fv
+  rm SpaceMono.zip
+  cd
+  else
+    echo "SpaceMono font already installed"
+fi
